apps/api/requirements.txt
# FastAPI framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Data processing
fastf1==3.3.0
pandas==2.2.0
numpy==1.26.4

# Caching
redis==5.0.1

# Database & Storage
supabase==2.3.4

# Auth
python-jose[cryptography]==3.3.0

# HTTP client
httpx==0.26.0

# Environment variables
python-dotenv==1.0.1

# Pydantic settings
pydantic-settings==2.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Type hints
typing-extensions==4.9.0
